+++ releasenotes.md
@@ -1,3 +1,4 @@
 Whats New in v10.8
 --------------------------
+- Under the hood improvements for Finished Product page
 - This release may include more changes, for full details please refer to the CHANGELOG